Add grid utility to crop unused rows and columns
